+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# next
+- push improvements. 
+- `Movement` mixin improvements.
+
 # 3.9.2
 - MiniMap improviments. Fix issue [#517](https://github.com/RafaelBarbosatec/bonfire/issues/517)
 - Raname `BouncingObject` to `ElasticCollision`.